Changes for handling Shortcuts with traitlets in consoleWidget
